sys-apps/hwids: add mapping for ChromeOS keyboards
Add mappings for "Hamburger" and "Assistant" keys on Hammer, Wand,
Whiskers, and Wallaby bases. Given that they are either external or
detachable and can be shared between devices it makes sense to keep the
mappings in a central place instead of keeping adding them to overlays.

BUG=b:112619632
TEST=Build and boot, check emitted keys with evtest

Change-Id: I3c43a6af754fa551ba55bb6ecf64ff47d2186edb
Reviewed-on: https://chromium-review.googlesource.com/1194661
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>