sys-devel/arc-llvm: Only install one copy of libLLVM
arc-llvm installs a copy of libLLVM.so into the build directory.  The
copy of the lib in build/ is the one that is used at build time by other
packages that link against arc-llvm (ie arc-mesa).

However, build_image only includes the contents of the vendor partition
in the final image, and thus we currently also make a copy into vendor/
for use at runtime.

Having two copies, however dramatically increases the size of the build,
however, since each 26 MB .so has its own 1 GB .debug file, so instead
replace the build/ copy with a symlink to the vendor/ copy.  Doing so
ensures that the .debug corresponds to the version actually installed in
the image.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=b:112313068
TEST=equery-grunt s arc-llvm
 => 4.52 GiB -> 3.63 GiB
TEST=equery-grunt f arc-llvm
 => includes:
  /opt/google/containers/android/build/lib/libLLVM-6.0.so -> ../../vendor/lib/libLLVM-6.0.so
  /opt/google/containers/android/vendor/lib/libLLVM-6.0.so
  /usr/lib/debug/opt/google/containers/android/vendor/lib/libLLVM-6.0.so.debug

Change-Id: I6e791df642d180ec823e0bdb04d1dbf5e478f4ea
Reviewed-on: https://chromium-review.googlesource.com/1198723
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>