chromeos-chrome: use 'strip' instead of 'eu-strip' with LLD.
Previously we changed to use eu-strip when using LLD to link chrome.
This is because using 'strip' corrupts the GNU_RELRO part of program
header.
However, using eu-strip makes the dump_syms crash and we could not
get the valid breakpad symbol files. Using 'strip' from binutils works
fine in this case. I verified that the error reported in the bug was
gone, probably it is becaues we upgraded LLD after the bug was reported.
We also remove the link flag -Wl,--no-rosegment because we added that in
the LLD wrapper already.

BUG=chromium:808262
TEST=USE="-gold lld" emerge-samus chromeos-chrome
     checked the output of
     readelf -l on stripped chrome binary, the GNU_RELRO part is
     still there.
     cros_generate_breakpad_symbols --board samus
      /build/samus/opt/google/chrome/chrome
     it generates valid symbol files.

Change-Id: I54f604d5e3b3be2a92335c8f8a83931ff997f465
Reviewed-on: https://chromium-review.googlesource.com/1200323
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>