sys-devel/llvm: add tools USE flag to enable building without tools
LLVM tools are huge and built by default.  Some users of llvm, such as
amdgpu mesa, do not require the huge suite of LLVM tools.
So, add a USE flag to enable not building them.

One tricky bit is that there is still one tool - llvm-config, which is
required even when all other tools are disabled.  However, this tool
is actually built for the *host*, not the target; the llvm ebuild already
has a special case for doing this through the function "build_host_tools".
This trick does an extra cmake configuration using mycmakeargs and then
manual builds llvm-tblgen and llvm-config for the host.  During this
step we must force LLVM_BUILD_TOOLS so that it will actually build
llvm-config.

The second tricky bit is that currently multilib_src_install(),
unconditionally tries installing llvm-config.  We change that to be
conditional on our new useflag so that the (target built) llvm-config's
are only installed if they are built.

Note: we can set the USE to default false for llvm-6.0.0, since it is
always cross compiled and used on targets that do not need tools.
For the host, we use llvm-7.0.0 which has a separate ebuild and installs
things a bit differently.

BUG=chromium:875286
TEST=USE=tools emerge-grunt llvm && equery-grunt s llvm
 * sys-devel/llvm-6.0.0-r3
         Total files : 1674
         Total size  : 11.44 GiB
TEST=USE=-tools emerge-grunt llvm && equery-grunt s llvm
         Total files : 1429
         Total size  : 3.32 GiB
TEST=in both cases:
  /build/grunt/usr/lib/llvm/bin/llvm-config-host --libdir --includedir
  /build/grunt/usr/lib/llvm/lib64
  /build/grunt/usr/lib/llvm/include

Change-Id: Ib7f382b3bf984ee14232adc8068796cdebf1fbe5
Reviewed-on: https://chromium-review.googlesource.com/1200450
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>