bluez: Fix mkfifo already exists
When bluetoothd crashes, it fails to restart because pre-start script
fails when trying to create the already existing /run/bluetooth/fifo
file. Creating this file should be guarded with "if not exists",
additionally this should be moved to bluetoothlog-upstart.conf instead
of bluez-upstart.conf because this has nothing to do with bluez, to
minimize any other possible regression that affects bluez.

BUG=chromium:882198
TEST=run "restart bluetoothd" and it should succeed
     run "start bluetoothlog", "kill $(pidof btmon)", check that log
     file is written to /var/log/bluetooth/log.gz

Change-Id: Ib37c87da7ef9b856dbd39bd84c8330c848b8b1d6
Reviewed-on: https://chromium-review.googlesource.com/1214824
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>