chromeos-cr50-scripts: Display message if leaving dev mode
RMA reset will reboot the system. If GBB flag does not force dev switch
on, the system will leave dev mode and become unable to boot from USB
shim. Display a message to tell the user to re-enter dev mode to boot
from USB again.

It will add 1~2 seconds before the user can enter the auth code due to
flashrom, but it should not be a problem because the user still needs
time to scan the QR code and get the auth code from the website.

BUG=b:111476604
TEST=manually test on DUT. RMA reset will show the message when GBB flag
     does not force dev switch on.

Change-Id: I38502a14263714c21413e81ab9ad2a109311299f
Reviewed-on: https://chromium-review.googlesource.com/1215510
Commit-Ready: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Nick Sanders <nsanders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>