Cherrypick fixes for llvm-next.
This CL picks up the fix in r340101, to fix a warnings issue we
are seeing with llvm-next in the kernel. It also updates an
existing patch, to apply cleanly to the llvm-next code.

BUG=chromium:883469
TEST=built llvm-next with patches. emerge-squawks kernel failed
without new cherrypick; passed with new cherrypick.

Change-Id: I710194280defbdbe60d3eb265e74b30fdaf2a86a
Reviewed-on: https://chromium-review.googlesource.com/1226390
Tested-by: Caroline Tice <cmtice@chromium.org>
Trybot-Ready: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>