profiles: chromeos: reset default compiler settings
The default CFLAGS under profiles/arch/ just uses -O2.  We already
add that in our default CFLAGS in make.conf.generic-target which
means we end up with -O2 -O2.  The same thing applies to CXXFLAGS.
For LDFLAGS, the arch/ default is -Wl,-O1, but we add -Wl,-O2, so
the flags have both which is pointless.

Lets simplify the base to remove this redundancy.  In practice,
there should be no difference because the accumulation is the same.

BUG=chromium:489641
TEST=precq passes
TEST=`portageq-$BOARD envvar CFLAGS CXXFLAGS LDFLAGS` is unchanged

Change-Id: I711bf4da9b23e229f594fb4e5f0ad80b83563935
Reviewed-on: https://chromium-review.googlesource.com/1226690
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>