make.conf.generic-target: rework default CFLAGS/CXXFLAGS
We declare a new BOARD_COMPILER_FLAGS variable which is a better
name than MARCH_TUNE so we can transition boards to it and eventually
drop MARCH_TUNE entirely.  This also lets us define it in the profile
instead of make.conf so we can drop make.conf in boards entirely.

We pull the common flags into a new COMMON_COMPILER_FLAGS variable.
Appending CFLAGS to CXXFLAGS leads to duplication (when the default
CFLAGS=-O2 CXXFLAGS=-O2, we end up with CXXFLAGS="-O2 -O2"), and it
prevents us from putting C-specific flags in CFLAGS.

BUG=chromium:489641
TEST=precq passes
TEST=`portageq-$BOARD envvar CFLAGS CXXFLAGS LDFLAGS` is unchanged

Change-Id: I75c39635e4cf9ac2c8af158dd94c02c440e4aa94
Reviewed-on: https://chromium-review.googlesource.com/1226691
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>