virtual/rust: Add virtual package for rust toolchain
Add a virtual package for the rust toolchain.  The version of this
package must match the version of the rust toolchain.  Rust packages
will depend on this virtual so that they will automatically be rebuilt
when we update the toolchain.

Also update the rust version in the package.provided file to match the
actual toolchain version.

BUG=none
TEST=`emerge-$BOARD virtual/rust`

Change-Id: I6e6c888d3585af97308405483319bc74eb740219
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1244593
Reviewed-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>