edk2: Install the tianocore image
At present we only install CBFS (Coreboot Filesystem) image in /firmware.
Install the actual tianocore image as well so that we can use this with
the altfw (alternative firmware) feature.

At some point we may wish to remove the CBFS option, but leave it for now.

BUG=chromium:837018
BRANCH=grunt
TEST=emerge-grunt edk2
$ ls /build/grunt/firmware/tianocore
UEFIPAYLOAD.fd

Change-Id: I85023bbe2759c54336d1954cb0aa5fb0564b91f7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1251881
Reviewed-by: Julius Werner <jwerner@chromium.org>