cros-kernel2: Add a 'debug' use flag for general purpose debug features
The kernel has many debug facilities that aren't turned on unless a
specific use flag is specified when building the kernel. Let's add a
generic 'debug' use flag that turns on all the debugging options so that
developers can easily turn on all the various kernel config options for
debugging. The assumption is that this use flag can and will degrade
performance significantly. We want to use this to catch kernel bugs in
test configurations, so this shouldn't be enabled in any sort of
production systems.

BUG=None
TEST=Build and boot kernel, modprobe configs && check /proc/config.gz
for all options enabled here modulo ones which are arch specific

Change-Id: Idc76931a9113003b3d57312e183f445dc7c9eec8
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1262158
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>