cr50-scripts: add check if board id should be set
Add check_device to cr50-set-board-id, so there is a standard way to
decide if setting the board id should be skipped. This will be used in
cr50-result and chromeos_postinst to determine if the board id should
be set.

This also preserves the cr50-result behavior to skip the reboot after
update in certain situations. check_device checks the device state and
exits with a non zero exit status if setting board id should be skipped.
It exits with a status of 4 if cr50-result should also skip the reboot.

These are the exact settings where setting board id or reboot will be
skipped.

Skip setting board id if mainfw_type is not normal or OS debug features
are enabled.

Skip board id and reboot if the flashrom command fails to run or if
SW WP is disabled.

BUG=b:115593945
BRANCH=none
TEST=verify cr50-result still sets board id the same way it did before.

Change-Id: Iea9cebf3ae584c7ca6fc02e185a2a1b6c7b22bac
Signed-off-by: Mary Ruthven <mruthven@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1294769
Commit-Ready: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>