selinux-policy: check non-tmpfs filetrans is defined in file_contexts
This will check filetrans_pattern in sepolicy, transferring from
nontmpfs, and to nontmpfs, is defined in file_contexts to reduce the
possibility an out-of-date file_contexts damages the file_contexts in
stateful parition during an upgrade.

This is not a prevention for such problem, since it only checks the new
label gets defined in the file_contexts in a loose way. It can't confirm
the file_contexts is exact match with filetrans_pattern.

BUG=b:116072767,b:114172200
TEST=emerge
CQ-DEPEND=CL:1286150

Change-Id: Ic30878ec90f40dac02341e1293da8118edc537de
Reviewed-on: https://chromium-review.googlesource.com/1286309
Commit-Ready: Qijiang Fan <fqj@google.com>
Tested-by: Qijiang Fan <fqj@google.com>
Reviewed-by: Kenny Root <kroot@google.com>