net-misc/rmtfs: don't keep FS in /boot
The /boot directory exists on test images, but not necessarily on base
images, so minijail's /var/lib/rmtfs/boot -> /boot bind mount won't work
(it's trying to create a /boot directory on the "minimilastic-mountns"
rootfs, which is read-only). Since we have the source rmtfs and because
/boot is a silly place for what essentially is a virtual, stateful
partition, let's move it to /var/lib/rmtfs/boot instead.

BUG=chromium:895056
TEST=`status rmtfs`, wifi on Cheza

Change-Id: I37492d960c6b02376aec883729a41f5013cabb1c
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1286357
Reviewed-by: Ben Chan <benchan@chromium.org>