vboot_reference: Change libzip dependency by sub slot
We are going to uprev libzip to 1.3.0, and since vboot_reference is part
of cros_sdk, we need to get the signing tools as well.

The new libzip needs sub slot, which is defined in EAPI=5, so we have to
change both.

BUG=chromium:875551
TEST=sudo emerge vboot_reference;
     emerge-$BOARD vboot_reference; build_image test
CQ-DEPEND=CL:1285990

Change-Id: I6f80ec7ab2470a5c02f2ed3638d0c9cacf52a8dc
Reviewed-on: https://chromium-review.googlesource.com/1287989
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>