cros-kernel2: Add xz use flag for compression choice and use it
The kernel is too big sometimes and so we'd like to shrink the size down
with xz. Add a use flag for this purpose and reuse it in the 'debug' use
flag for simplification. This way debugging kernel images don't get too
big on x86/amd64 images and we can use this to shrink 'ubsan' built
images too.

BUG=chromium:894237
TEST=Build kernel image and see if it fits into 16MB
Signed-off-by: Stephen Boyd <swboyd@chromium.org>

Change-Id: I996fa649e48a8529ab48366db9b3fb600b62152c
Reviewed-on: https://chromium-review.googlesource.com/1291650
Commit-Ready: Stephen Boyd <swboyd@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Yu Zhao <yuzhao@chromium.org>