chromeos-bootimage: Drop legacy payload logic
This is not used anymore. We should use altfw instead which is more
flexible:

   - It allows multiple bootloaders
   - It provides a menu for the user

It also provides a path to the word 'legacy' which is clearly a misnomer
since some bootloaders are still widely used.

The signing feature appears to not be implemented. It can be implemented
using altfw instead.

BUG=chromium:895195
TEST=emerge-grunt chromeos-bootimage

Change-Id: Ia9e7cc6f7cbc4330c0172e4725b8777152604c73
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1292995
Commit-Ready: Duncan Laurie <dlaurie@google.com>
Tested-by: Duncan Laurie <dlaurie@google.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>