shellcheck: Allow local in POSIX scripts
Officially, local isn't allowed in POSIX shell scripts, so shellcheck
warns about it under SC2039.  All the shells we use actually allow it and
we make use of it, so we don't want to warn about it.  An upstream
discussion was started at https://github.com/koalaman/shellcheck/issues/1341,
but until that's resolved, we can locally patch out the check.

BUG=chromium:891826
TEST=shellcheck src/platform2/init/dev_utils.sh

Change-Id: I6b68f98862052facf868a81f645c827ab551860a
Reviewed-on: https://chromium-review.googlesource.com/1295291
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>