cros-kernel2: Rename 'usban' to 'kernel_ubsan'
The ubsan builder can't fit the kernel image into the 16MB partition
that we have by default. Until that is fixed, let's rename this flag to
kernel_ubsan so the ubsan builder is happy. In the meantime, we can
increase the size and put this back to ubsan when it's known to be
working.

BUG=chromium:894237
TEST=Build amd64-ubsan image and see kernel fits into partition

Change-Id: I51541e5aa6fefeece5a5e7744775cc5a81b09e9e
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1296034
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>