Change the dependency of system-locales.
The system-locales requires exactly the same version as glibc. However,
we could not upgrade glibc and system-locales in the same time. This is
because the glibc is built by the sdk bot, but the system-locales is
built for the target. If we upgrade these two package at the same time,
in the commit queue buildbot, the buildbot will try to build glibc for
target and it will fail. So glibc and system-locales can fall out of
sync during an upgrade.

This also make system-locales work even glibc version is greater than 2.24.
We can work on a better ebuild file after glibc lands.

BUG=chromium:834385
TEST=cbuildbot

Change-Id: I705b5d63a1cec0634ec2468bf9acac7cc82874d5
Reviewed-on: https://chromium-review.googlesource.com/1299576
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>