virtual/target-chromium-os-test: temporarily disable toolchain-tests.
Current platform_ToolchainTest only works for glibc 2.23. CL:1139076
only works for glibc 2.27. And we could not upgrade glibc and this test
 in the same time. This is because the glibc is built by the sdk bot,
but the test is for the target. If we upgrade these two package at the
same time, in the commit queue buildbot it will cause unittest fail.
To workaround that, we disable this test temporarily and check in the
change for glibc 2.27. After glibc 2.27 roll finishes, we enable this
test again.

BUG=chromium:834384
TEST=toolchain-tests does not run in cros_run_unittest.

Change-Id: Icc6e57353746571f7d00ec2696242db84d7b36ff
Reviewed-on: https://chromium-review.googlesource.com/1300194
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>