ebuild/chromeos-ec: fix main firmware install
this fixes a bug where in the case where we build multipe EC firmware
blobs (EC_FIRMWARE_EXTRA=...), we end up grabbing the build output
from the last board in the list of boards and install it as the
primary (/build/${BOARD}/firmware/ec.bin).

EC_BOARD_NAMES seems to be alpha-sorted, so this is still not quite
the right thing to do.  we should probably just drop this bit of code
and fix anything that relies on this behavior to use the unibuild
logic for accessing ec.bin.

BUG=b:118463372
TEST=/build/${BOARD}/firmware/ec.bin is now the main EC firmware

Change-Id: I659fa4f019a223452b74a44888dccc228d63a26d
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1300654
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: caveh jalali <caveh@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>