ubsan: Compress kernel image as much as possible
Let's use the XZ compression method for the kernel when we're building
an ubsan image. Ubsan in the kernel makes the vmlinuz quite large so
compressing it with a slower compression method allows it to fit into
a smaller partition size while sacrificing some decompression time, but
this is all debug anyway so it seems to be a worthwhile tradeoff.

BUG=chromium:894237
TEST=Build ubsan profile and see kernel compressed with xz

Change-Id: I3304f7f85488bba84864d6bd92bd31443c54a1c5
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1318403
Reviewed-by: Douglas Anderson <dianders@chromium.org>