gcc: allow us to force -Wno-error in builds
For testing LLVM/clang nearer to ToT, it's useful if new warnings aren't
fatal. This allows us to set a special environment variable to get a
final say in the fatalness of warnings.

Tweaking CFLAGS/CXXFLAGS isn't sufficient because we want warnings to
never be fatal for these builds, and it's trivial for a single ebuild to
place its own -Werror (or similar) after the user-provided
CFLAGS/CXXFLAGS. This behavior is desirable for things that rely on
-Werror like autotools, but not great in the case of projects like
Chrome or things in platform2.

More importantly, we have plans for building extra tooling on top of
this support. Please see bug 914081 for more about that.

BUG=chromium:903652, chromium:914081
TEST=emerged cross-gcc and built packages that should've otherwise had
     fatal warnings. Needed a separate hook to actually `export` the
     variable, naturally.

Change-Id: I0bd794c9c06913410722e94b04803cdc7b461bf2
Reviewed-on: https://chromium-review.googlesource.com/1328150
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>