llvm: refactor EGIT_REPO_URIS + do full ToT syncs
This CL refactors to do a full git pull/etc when USEing llvm-tot.
This is necessary because our epatch_after/epatch_before magic only
works when we have .git dirs, and the current pulling logic won't
give those to us, apparently.

The simplest way to do that seems to be to refactor a bit.

BUG=chromium:903652
TEST=emerged llvm (well, got to the 'apply patches' step) with
     llvm-next, llvm-tot, and llvm

Change-Id: Id208d3fda2d760643633e044c8dbc4520272c39b
Reviewed-on: https://chromium-review.googlesource.com/1328156
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>