chromeos-bootimage: add altfw payload hashing
Under certain circumstances, we would like to allow booting
only Google-issued RW_LEGACY firmware.  In order to allow for
the option to verify the payload being booted, we calculate
and store SHA-256 hashes of each payload in AP-RW CBFS, which
is signed and rollback-protected already as part of AP-RW.

A hash of each payload will be stored in the file:

  altfw_hash/<name>.sha256

Originally written prior to altfw as CL:1169027.

BUG=b:110721285,b:111577108,b:119384389,b:118786884,chromium:837018
TEST=USE="altfw tianocore" emerge-eve chromeos-bootimage
     cbfstool image.bin print -r FW_MAIN_A
     check for altfw_hash/tianocore.sha256
CQ-DEPEND=CL:1341364

Change-Id: I5f20543e245fe7c5e0b142f81346275cf1623fd6
Reviewed-on: https://chromium-review.googlesource.com/1331272
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>