llvm: Disable -faddrsig option for host builds.
clang upstream now enables -faddrsig option as default
which produces sections that binutils strip does not understand.
This causes strip to fail for some bjects and leave a bunch of stray
object files installed to the chroot.

This option is to efficiently enable "-icf=all" linker option but
there is no use of "-icf=all" in the host builds. So let's drop this
option from the host wrapper for now.

Should also evaluate dropping this from target builds except chrome
but .o files are automatically masked from being installed on image
so this isn't a problem yet for target packages.

BUG=chromium:894723
TEST=sudo emerge libbrillo and checked that no .o files are installed.

Change-Id: I2f0ab80eeeaa6f405ad5bbda6e3c34e33e00cc8c
Reviewed-on: https://chromium-review.googlesource.com/1338498
Commit-Ready: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>