dev-lang/python: Update Python 2.7 distutils to not libdir on user
We set libdirname for cross-compilation. The code currently only adds
this variable to the class attributes when not self.user and self.home
is not None, yet we reference this attribute later without testing for
the same conditionals. This adds these conditionals before performing
this look-up.

This fixes `pip install --user` inside the chroot.

BUG=none
TEST=emerge python:2.7; pip install --user setuptools

Change-Id: Ib8397ed397d8c079747556d140bc19d6c91bbde7
Reviewed-on: https://chromium-review.googlesource.com/1339061
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>