arc-llvm: Attempt to reduce build times
arc-llvm takes so long to rebuild that it can cause timeouts on paladin
runs.  Ideally it shouldn't be getting rebuilt except when we change it,
but the current long chain of portage deps triggers rebuilds relatively
often.  As a short-term improvement, three changes that will help reduce
the arc-llvm build times:

* Disable building libLTO, since we don't use it.
* Disable building the runtimes, since we don't use any of them.
* Patch out installation of static libraries, since dependent packages
  only use the dynamic library.

The combination of these reduces the package size by ~2GB.  On an idle
workstation, it reduces compile time by 15-20%.  The effect on a heavily
loaded build machine is harder to determine, but it should help a
little.

BUG=chromium:906078
TEST=./build_packages --board=grunt

Change-Id: Ic93efc65e4c2507c7489182c136af73372600aa8
Reviewed-on: https://chromium-review.googlesource.com/1344373
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>