dlcservice: Add dlcservice_util to dlcservice ebuild.
Add dlcservice_util to dlcservice ebuild. It will be installed to
/usr/local/bin/. We want to avoid adding a new ebuild for the
dlcservice_util, but in order to add it to the dlcservice ebuild we
have to get rid of the runtime dependency of dlcservice-client on
dlcservice and add a compile-time dependency on dlcservice-client to
dlcservice. This is a similar pattern to what is used in
update_engine.

BUG=chromium:852171
TEST=Build dlcservice, client utility is installed to /usr/local/bin/.
CQ-DEPEND=CL:1351400

Change-Id: I06cf4da19e752de7297ceb546e72bec9648e98c6
Reviewed-on: https://chromium-review.googlesource.com/1347203
Commit-Ready: Colin Howes <chowes@google.com>
Tested-by: Colin Howes <chowes@google.com>
Reviewed-by: Amin Hassani <ahassani@chromium.org>