u-boot: Add support for building sandbox
U-Boot's sandbox is a convenient way to build and test vboot
functionality. Add a USE flag to build this. This requires SDL so add that
to the sdk.

This works by building using the chromeos_sandbox target, using the host
toolchain (not cross-compiled). The resulting output is U-Boot along with
SPL and TPL. There is also a fully ready Chromium OS firmware image which
is installed, for testing purposes.

Instructions for running the result are in the ebuild.

The USE_STDINT option is not needed with vboot anymore, so drop it.

BUG=chromium:908228
TEST=USE="sandbox vboot" sudo -E emerge u-boot
See that it builds OK
Follow instructions at the top of the ebuild to run U-Boot inside and
outside the chroot.

Change-Id: Iaffe6d4ccccc2514cf6efbb793356eef2973a2b2
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1350223
Reviewed-by: Mike Frysinger <vapier@chromium.org>