chromeos-base/portier: add shill to CROS_WORKON_SUBTREE
Portier needs access to the shill sources during build. This fixes
sandbox exceptions when building portier.

This ideally should not be necessary, since shill headers are installed
in /usr/include, but due to https://crbug.com/809389, shill's source
directory is earlier in the #include search path, which breaks the
sandbox.  Adding it to CROS_WORKON_SUBTREE seems like the least bad way
to work around this until the underlying bug is fixed.

BUG=chromium:809389
TEST=emerge-nami portier

Change-Id: I710647c3ab83e96964ce300a958f3f445563138d
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1351094
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>