mosys: enable meson testing
This will try to run any meson tests defined for mosys, which is
currently the empty set. The three unit tests in the current repo
are old and do not compile.

Next steps will be to convert these tests from cmockery to cmocka
and then enable them.

BUG=chromium:910641
TEST=Run local builds and tests for X86 and ARM

Change-Id: I160d4f06c2e82a2d0dae9c36937d131dfc25d2e0
Reviewed-on: https://chromium-review.googlesource.com/1357121
Commit-Ready: Gregory Meinke <gmeinke@chromium.org>
Tested-by: Gregory Meinke <gmeinke@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>