libchrome: Uprev base/json to r576297
This CL uprevs to base/json version to r576297 to catch important
security hardening work.

It is done this way instead of checking the code directly into the
libchrome repo so it doesn't interfere with upcoming changes to
libchrome. It is not done as a regular patch to simplify code review.

r576297 was picked since it is planned to be the next uprev target of
libchrome, so it's going to be easy to handle (basically, just revert
this CL). Moreover, it contains all hardening CLs except one
(CL:1308196).

To simplify code review,
- PS1 contains vanilla r576297 code and
- PS2+ contains the changes to make it compile with libchrome again.

There were some minor changes to make the API backwards compatible:
- Adding back JSON_DETACHABLE_CHILDREN (does not do anything, but some
  Chrome OS code still uses the flag).
- const StringPiece& json instead of StringPiece in ReadAndReturnError.
Note that when recompiling dependencies, the build system does not pick
up incremental changes from a libchrome-456626-rN.ebuild ->
libchrome-456626-rN+1.ebuild uprev.

BUG=chromium:860181
TEST=Manually ran all unit tests and fuzzer tests.

Change-Id: I79529fc297c0b62cb9100bedb840c894b7851354
Reviewed-on: https://chromium-review.googlesource.com/1349354
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>