dev-lang/rust: upgrade to rust 1.31
This version is the first with the Rust 2018 edition. Yay!

The only other thing of note about this upgrade is that the miri tool
had to be disabled because the build for it is broken. An updated
version of Rust will likely clear up that issue.

The recently enabled sanitizers patch also needed to be rebased.

BUG=None
TEST=sudo emerge rust
CQ-DEPEND=CL:1371686

Change-Id: Ie32193c83ae01a3b905b736b3ea6169c1e7a0ccc
Reviewed-on: https://chromium-review.googlesource.com/1366446
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>