chromeos-bootimage: Support installing EC image from wilco
Add a USE flag for wilco_ec and use the same path as cros_ec to add
it to the firmware image.

The Wilco EC also needs a version file added to CBFS so if this USE
flag is selected it will also add that file.

BUG=b:120628933
TEST=emerge-sarien chromeos-bootimage and verify with cbfstool:

ecrw          0xfc000    raw     184192 none
ecrw.hash     0x8480     raw         32 none
ecrw.version  0x8500     raw          3 none

Change-Id: Ia639672cf6b464ca701e775b6901a8adc5470770
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1367126
Reviewed-by: Simon Glass <sjg@chromium.org>