ectool: Add separate USE flag for ectool
In order to support systems without ectool we need a flag that
explicitly indicates whether or not it should be present.  To
do this I split out the ectool build/install process and only
call those functions if the USE flag is enabled.

By default this flag is true, and it is disabled only on systems
that do not use the Chromium EC.

The new flag is also used by the initramfs ebuild to provide an
indication to the initramfs build system so it can include ectool
only if it is present instead of always expecting it to exist.

BUG=chromium:913766
TEST="USE=factory_netboot_ramfs emerge-sarien chromeos-initramfs"
CQ-DEPEND=CL:1372266

Change-Id: Ia93ff56f07b23465d586cf9b5f7ac0d5b227a246
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1372284
Reviewed-by: Simon Glass <sjg@chromium.org>