shellcheck: Add predefined ebuild/eclass variables
This adds the internal portage ebuild/eclass variables to the list of
known variables so that SC2034 isn't printed for variables that appear
to be unused.  It was generated by:

1. Add the list of variables from
   https://devmanual.gentoo.org/eclass-reference/ebuild/index.html
2. Use get_vars.py to extract all the @VARIABLE and @ECLASS-VARIABLE
   declarations from eclasses under chromiumos-overlay and
   portage-stable.
3. Run the new shellcheck on all ebuilds under chromiumos-overlay and
   portage stable and add a few stragglers that were not declared
   in the expected formats.  I only added variables used at least 10
   times; the remaining uses can be annotated individually if we decide
   they're real.
4. Add a few of the variables to the space-free and array lists.  This
   part is definitely incomplete, but covers a few of the common ones.

The downside to this approach is that an ebuild can use variables from
any eclass, not just the ones that it inherits.  Fixing this properly
requires following the chains of eclass inheritance.  Since the current
setup is better than today's situation of entirely disabling SC2034 on
ebuilds, I've saved that for follow-up work.

BUG=chromium:898580
TEST=shellcheck all ebuilds and check SC2034 errors

Change-Id: I81375b469f053ac472b48353fd61b0857e5c7104
Reviewed-on: https://chromium-review.googlesource.com/1373917
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>