kmod: disable support for /run/modprobe.d
We don't have any packages that write module fragments to this path,
so lets stop reading it.  This avoids attacks where someone can write
to this path and then trigger a module load event later on (and the
kernel will run it in the root namespace as root).

BUG=chromium:780039
TEST=precq passes
CQ-DEPEND=CL:1380992

Change-Id: Ida0c7879828029689f3cf3c55014d74ebb7bc7c2
Reviewed-on: https://chromium-review.googlesource.com/1380773
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>