chromeos-base: Don't specify 'kernel.panic = -1' in 00-sysctl.conf
We already specify 'panic=-1' on the kernel command line in
'build_kernel_image.sh'.  Having it _also_ in 00-sysctl.conf just
makes it harder for developer to override (like if they want to drop
into kgdb on panic) without any actual benefit.  Specifically if we
had to wait all the way for userspace to set this up then we'd already
be failing to reboot on panic for any panics that happened before
userspace started up.

Let's remove this.  Now developers that want to enable kgdb can just
add "panic=0" to the end of their kernel command line and override the
default one from build_image.

CQ-DEPEND=CL:1382881, CL:1383232
BUG=chromium:908726
TEST=tast -verbose run ${IP} kernel.Limits passes unless I override

Change-Id: I6ad7ae1626e31a86636d00fdb8c37cd903a010b3
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1382879
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>