chromeos-chrome: Build capture_unittests and video_decode_accelerator_tests
Build capture_unittests and video_decode_accelerator_tests to
run it from a Tast test.
capture_unittests is a binary test for video capture stack
and it should be run on Chromium CQ and PFQ.
video_decode_accelerator_tests is a new video acceleartion
tests that will replace video_decode_accelerator_unittest
in the future.

BUG=chromium:893501, chromium:905954
TEST=locally build with CL:1375210 for amd64-generic

Change-Id: Ifd83d7c99f9595bbdf3da6f407d47c5c7d7493f8
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1383771