Skip tests for some ebuilds that fail.
libprotobuf-mutator, openssl, and protobuf are packages for third party
code that do not normally run tests. This change skips tests for these
ebuilds so they can be included as dependencies of tests run by first
party ebuilds. A rev bump was not performed because there are no changes
to the installed files.

BUG=None
TEST=emerge-${BOARD} chromeos-ec && FEATURES=test emerge-${BOARD} \
  chromeos-ec

Change-Id: Iccd05ab5c9526cd8c12f4a93c864c3b9e373e7f0
Reviewed-on: https://chromium-review.googlesource.com/1384354
Commit-Ready: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>