chromeos-chrome: Add test in ebuild to make sure hot symbols are reordered
This patch add tests before installation in Chrome ebuild to make Chrome binary
puts hot symbols in a section .text.hot, which is placed before .text. Errors
will be printed if not and installation will be stopped.

This is also a test to make sure
https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1378594/
is applied correctly on Chrome.

The check needs to be enabled with USE flag "strict_toolchain_checks".

TEST=Tested on caroline
BUG=chromium:912781, chromium:912780

Change-Id: I0c7b6794a7e5574894c744c3be48c7e42e627eb5
Reviewed-on: https://chromium-review.googlesource.com/1394123
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>